gainer module added
